% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{bl_gestage_pt_centile}
\alias{bl_gestage_pt_centile}
\title{International Newborn Size References for Very Preterm Infants: table of
centiles for birth length-for-gestational age for very preterm infants}
\format{
A tibble with 882 rows and 5 columns:
\describe{
\item{\code{week}}{Gestational age weeks component}
\item{\code{day}}{Gestational age days component}
\item{\code{sex}}{Sex; 1 = male; 2 = female}
\item{\code{centile}}{Centile - 3rd, 5th, 10th, 50th, 90th, 95th, 97th}
\item{\code{bl}}{Birth length in centimetres}
}
}
\source{
Villar J, Giuliani F, Fenton TR, Ohuma EO, Ismail LC,
Kennedy SH et al. INTERGROWTH-21st very preterm size at birth reference
charts. Lancet 2016, 387(10021):844-45.
}
\usage{
bl_gestage_pt_centile
}
\description{
International Newborn Size References for Very Preterm Infants: table of
centiles for birth length-for-gestational age for very preterm infants
}
\examples{
bl_gestage_pt_centile

}
\keyword{datasets}
