% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/03-calculate_fetal_growth.R
\name{calculate_ac_gestage}
\alias{calculate_ac_gestage}
\title{Calculate abdominal circumference for gestational age centiles and z-scores}
\usage{
calculate_ac_gestage(gestage = NULL, ac = NULL)
}
\arguments{
\item{gestage}{Gestational age. Gestational age can be provided as a string
with a format of weeks+days or as numeric of exact weeks in decimal format}

\item{ac}{Abdominal circumference in millimetres}
}
\value{
A list of 2 elements containing 1) abdominal circumference-
for-gestational age z-score; and, 2) abdominal circumference-
for-gestational age centile.
}
\description{
Calculate abdominal circumference for gestational age centiles and z-scores
}
\examples{
calculate_ac_gestage(gestage = "78+0", ac = 55)

}
