% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/03-calculate_fetal_growth.R
\name{calculate_bpd_gestage}
\alias{calculate_bpd_gestage}
\title{Calculate bi-parietal diameter for gestational age centiles and z-scores}
\usage{
calculate_bpd_gestage(gestage = NULL, bpd = NULL)
}
\arguments{
\item{gestage}{Gestational age. Gestational age can be provided as a string
with a format of weeks+days or as numeric of exact weeks in decimal format}

\item{bpd}{Bi-parietal diameter in millimetres}
}
\value{
A list of 2 elements containing 1) bi-parietal
diameter-for-gestational age z-score; and, 2) bi-parietal
diameter-for-gestational age centile.
}
\description{
Calculate bi-parietal diameter for gestational age centiles and z-scores
}
\examples{
calculate_bpd_gestage(gestage = "78+0", bpd = 20)

}
