% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/03-calculate_fetal_growth.R
\name{calculate_efw}
\alias{calculate_efw}
\title{Calculate estimated fetal weight given abdominal circumference and head
circumference}
\usage{
calculate_efw(ac = NULL, hc = NULL)
}
\arguments{
\item{ac}{Abdominal circumference in centimetres. Use measurements made
between 22+0 and 40+0 gestational weeks only.}

\item{hc}{Head circumference in centimetres. Use measurements made between
22+0 and 40+0 gestational weeks only.}
}
\value{
Estimated fetal weight in grams.
}
\description{
Calculate estimated fetal weight given abdominal circumference and head
circumference
}
\examples{
calculate_efw(ac = 55, hc = 90)

}
