% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/03-calculate_fetal_growth.R
\name{calculate_efw_gestage}
\alias{calculate_efw_gestage}
\title{Calculate estimated fetal weight for gestational age centiles and z-scores}
\usage{
calculate_efw_gestage(gestage = NULL, ac = NULL, hc = NULL)
}
\arguments{
\item{gestage}{Gestational age. Gestational age can be provided as a string
with a format of weeks+days or as numeric of exact weeks in decimal format}

\item{ac}{Abdominal circumference in centimetres}

\item{hc}{Head circumference in centimetres}
}
\value{
A list of 6 elements containing 1) estimated fetal weight in grams
  based on abdominal circumference and head circumference; 2) skewness of
  gestational age; 3) mean gestational age; 4) coefficient of variation of
  gestational age; 5) estimated fetal weight-for-gestational age z-score;
  and, 6) estimated fetal weight-for-gestational age centile.
}
\description{
Calculate estimated fetal weight for gestational age centiles and z-scores
}
\examples{
calculate_efw_gestage(gestage = "30+0", ac = 26, hc = 29)

}
