% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/03-calculate_fetal_growth.R
\name{calculate_fl_gestage}
\alias{calculate_fl_gestage}
\title{Calculate femur length for gestational age centiles and z-scores}
\usage{
calculate_fl_gestage(gestage = NULL, fl = NULL)
}
\arguments{
\item{gestage}{Gestational age. Gestational age can be provided as a string
with a format of weeks+days or as numeric of exact weeks in decimal format}

\item{fl}{Femur length in millimetres}
}
\value{
A list of 2 elements containing 1) femur length-for-gestational age
  z-score; and, 2) femur length-for-gestational age centile.
}
\description{
Calculate femur length for gestational age centiles and z-scores
}
\examples{
calculate_fl_gestage(gestage = "78+0", fl = 10)

}
