% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/01-calculate_gestage.R
\name{calculate_gestage}
\alias{calculate_gestage}
\title{Calculate gestational age given crown-rump length and/or head circumference
and femur length based on ultrasound}
\usage{
calculate_gestage(date = Sys.Date(), crl = NULL, hc = NULL, fl = NULL)
}
\arguments{
\item{date}{Date when measurement/s was/were taken. Defaults to current
date.}

\item{crl}{Crown-rump length measurement in millimetres}

\item{hc}{Head circumference measurement in millimetres}

\item{fl}{Femur length measurement in millimetres}
}
\value{
A list of 3 elements containing 1) gestational age in weeks (numeric);
2) gestational age in weeks and days (character); and, 3) expected date
of delivery (date).
}
\description{
Calculate gestational age given crown-rump length and/or head circumference
and femur length based on ultrasound
}
\examples{
calculate_gestage(date = "2016-23-12", crl = 17)
calculate_gestage(date = "2016-23-12", hc = 130, fl = 20)

}
