% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/04-calculate_gest_weight.R
\name{calculate_gwg_gestage}
\alias{calculate_gwg_gestage}
\title{Calculate gestational weight gain-for-gestational age z-score and centile}
\usage{
calculate_gwg_gestage(gestage = NULL, baseline = NULL, current = NULL)
}
\arguments{
\item{gestage}{Gestational age. Gestational age can be provided as a string
with a format of weeks+days or as numeric of exact weeks in decimal format}

\item{baseline}{Baseline weight in kilograms (up to 2 decimal places) in the
first trimester of pregnancy (less than 14 weeks gestational age)}

\item{current}{Weight at current visit in kilograms (up to 2 decimal places)}
}
\value{
A list of 3 elements containing 1) gestational weight gain in
kilograms; 2) gestational weight gain-for-gestational age z-score; and,
3) gestational weight gain-for-gestational age centile.
}
\description{
Calculate gestational weight gain-for-gestational age z-score and centile
}
\examples{
calculate_gwg_gestage(gestage = "32+0", baseline = 60, current = 69.52)

}
