% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/03-calculate_fetal_growth.R
\name{calculate_hc_gestage}
\alias{calculate_hc_gestage}
\title{Calculate head circumference for gestational age centiles and z-scores}
\usage{
calculate_hc_gestage(gestage = NULL, hc = NULL)
}
\arguments{
\item{gestage}{Gestational age. Gestational age can be provided as a string
with a format of weeks+days or as numeric of exact weeks in decimal format}

\item{hc}{Head circumference in millimetres}
}
\value{
A list of 2 elements containing 1) head circumference-for-gestational
  age z-score; and, 2) head circumference-for-gestational age centile.
}
\description{
Calculate head circumference for gestational age centiles and z-scores
}
\examples{
calculate_hc_gestage(gestage = "78+0", hc = 90)

}
