% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/03-calculate_fetal_growth.R
\name{calculate_ofd_gestage}
\alias{calculate_ofd_gestage}
\title{Calculate occipito-frontal diameter for gestational age centiles and z-scores}
\usage{
calculate_ofd_gestage(gestage = NULL, ofd = NULL)
}
\arguments{
\item{gestage}{Gestational age. Gestational age can be provided as a string
with a format of weeks+days or as numeric of exact weeks in decimal format}

\item{ofd}{Occipito-frontal diameter in millimetres}
}
\value{
A list of 2 elements containing 1) occipito-frontal
  diameter-for-gestational age z-score; and, 2) occipito-frontal
  diameter-for-gestational age centile.
}
\description{
Calculate occipito-frontal diameter for gestational age centiles and z-scores
}
\examples{
calculate_ofd_gestage(gestage = "78+0", ofd = 25)

}
