% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/02-calculate_sfh_gestage.R
\name{calculate_sfh_gestage}
\alias{calculate_sfh_gestage}
\title{Calculate symphysis-fundal height-for-gestational age z-score and centile}
\usage{
calculate_sfh_gestage(gestage = NULL, sfh = NULL)
}
\arguments{
\item{gestage}{Gestational age. Gestational age can be provided as a string
with a format of weeks+days or as numeric of exact weeks in decimal format}

\item{sfh}{Symphysis-fundal height measurement in centimetres up to 1
decimal place}
}
\value{
A list of 2 elements containing 1) symphysis-fundal
height-for-gestational age z-score; and, 2) symphysis-fundal
height-for-gestational age centile.
}
\description{
Calculate symphysis-fundal height-for-gestational age z-score and centile
}
\examples{
calculate_sfh_gestage(gestage = "40+1", sfh = 40)

}
