% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{crl_gestage_centile}
\alias{crl_gestage_centile}
\title{International fetal size standards in early pregnancy: table of centiles of
crown-rump length based on gestation age in weeks and days}
\format{
A tibble with 336 rows and 4 columns:
\describe{
\item{\code{week}}{Gestational age - weeks}
\item{\code{day}}{Gestation age - days}
\item{\code{centile}}{Centiles (3rd, 5th, 10th, 50th, 90th, 95th, 97th)}
\item{\code{crl}}{Crown-rump length in millimetres}
}
}
\source{
Napolitano R, Dhami J, Ohuma EO, Ioannou C, Conde-Agudelo A,
Kennedy SH et al. Pregnancy dating by fetal crown-rump length: a
systematic review of charts. BJOG 2014, 121(5):556-5.
}
\usage{
crl_gestage_centile
}
\description{
International fetal size standards in early pregnancy: table of centiles of
crown-rump length based on gestation age in weeks and days
}
\examples{
crl_gestage_centile

}
\keyword{datasets}
