% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{gestage_crl}
\alias{gestage_crl}
\title{International dating standards in early pregnancy: table of centiles of
gestational age in weeks and days based on crown-rump length in millimetres}
\format{
A tibble with 810 rows and 4 columns:
\describe{
  \item{\code{crl}}{Crown-rump length in millimetres}
  \item{\code{centile}}{Centile (3rd, 10th, 50th, 90th, 97th)}
  \item{\code{unit}}{Time unit for age value (week or day)}
  \item{\code{age}}{Age value}
}
}
\source{
Napolitano R, Dhami J, Ohuma EO, Ioannou C, Conde-Agudelo A,
    Kennedy SH et al. Pregnancy dating by fetal crown-rump length: a
    systematic review of charts. BJOG 2014, 121(5):556-5.
}
\usage{
gestage_crl
}
\description{
International dating standards in early pregnancy: table of centiles of
gestational age in weeks and days based on crown-rump length in millimetres
}
\examples{
gestage_crl

}
\keyword{datasets}
