% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{gestwt_gestage_centile}
\alias{gestwt_gestage_centile}
\title{International gestational weight gain standards: table of centiles of
gestational weight gain for women with normal body mass index based on
gestational age in exact weeks}
\format{
A tibble with 182 rows and 3 columns:
\describe{
  \item{\code{week}}{Gestational age in exact weeks}
  \item{\code{centile}}{Centiles (3rd, 5th, 10th, 50th, 90th, 95th, 97th)}
  \item{\code{gestwt}}{Gestational weight gain in kilograms}
}
}
\source{
Cheikh Ismail L, Bishop DC, Pang R, Ohuma EO, Kac G, Abrams B et al.
  Gestational weight gain standards based on women enrolled in the Fetal
  Growth Longitudinal Study of the INTERGROWTH-21st Project: a prospective
  longitudinal cohort study. BMJ 2016, 352(i):555.
}
\usage{
gestwt_gestage_centile
}
\description{
International gestational weight gain standards: table of centiles of
gestational weight gain for women with normal body mass index based on
gestational age in exact weeks
}
\examples{
gestwt_gestage_centile

}
\keyword{datasets}
