% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{gwg_gestage_sd}
\alias{gwg_gestage_sd}
\title{International gestational weight gain standards: table of z-scores of
gestational weight gain for women with normal body mass index based on
gestational age in exact weeks}
\format{
A tibble with 182 rows and 3 columns:
\describe{
  \item{\code{week}}{Gestational age in exact weeks}
  \item{\code{sd}}{z-scores (-3SD, -2SD, -1SD, 0, 1SD, 2SD, 3SD)}
  \item{\code{gestwt}}{Gestational weight gain in kilograms}
}
}
\source{
Cheikh Ismail L, Bishop DC, Pang R, Ohuma EO, Kac G, Abrams B et al.
  Gestational weight gain standards based on women enrolled in the Fetal
  Growth Longitudinal Study of the INTERGROWTH-21st Project: a prospective
  longitudinal cohort study. BMJ 2016, 352(i):555.
}
\usage{
gwg_gestage_sd
}
\description{
International gestational weight gain standards: table of z-scores of
gestational weight gain for women with normal body mass index based on
gestational age in exact weeks
}
\examples{
gwg_gestage_sd

}
\keyword{datasets}
