% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{hc_gestage_centile}
\alias{hc_gestage_centile}
\title{International fetal growth standards: table of centiles of head circumference
based on gestation age in exact weeks}
\format{
A tibble with 189 rows and 3 columns:
\describe{
  \item{\code{week}}{Gestational age in exact weeks}
  \item{\code{centile}}{Centiles (3rd, 5th, 10th, 50th, 90th, 95th, 97th)}
  \item{\code{hc}}{Head circumference in millimetres}
}

A tibble with 882 rows and 5 columns:
\describe{
  \item{\code{week}}{Gestational age weeks component}
  \item{\code{day}}{Gestational age days component}
  \item{\code{sex}}{Sex; 1 = male; 2 = female}
  \item{\code{centile}}{Centile - 3rd, 5th, 10th, 50th, 90th, 95th, 97th}
  \item{\code{hc}}{Head circumference in centimetres}
}
}
\source{
Papageorghiou AT, Ohuma EO, Altman DG, Todros T, Cheikh Ismail L,
  Lambert A et al. International standards for fetal growth based on serial
  ultrasound measurements: the Fetal Growth Longitudinal Study of the
  INTERGROWTH-21st Project. Lancet 2014, 384(9946):869-79.

Villar J, Giuliani F, Fenton TR, Ohuma EO, Ismail LC,
  Kennedy SH et al. INTERGROWTH-21st very preterm size at birth reference
  charts. Lancet 2016, 387(10021):844-45.
}
\usage{
hc_gestage_centile

hc_gestage_centile
}
\description{
International fetal growth standards: table of centiles of head circumference
based on gestation age in exact weeks

International Newborn Size References for Very Preterm Infants: table of
centiles for head circumference-for-gestational age for very preterm infants
}
\examples{
hc_gestage_centile

hc_gestage_centile

}
\keyword{datasets}
