% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/intergrowth.R
\docType{package}
\name{intergrowth}
\alias{intergrowth}
\alias{intergrowth-package}
\title{The INTERGROWTH-21st Project is a multi-centre, multi-ethnic,
population-based project, conducted between 2009 and 2014, in eight
demarcated urban areas: Pelotas, Brazil; Shunyi County, Beijing, China;
Central Nagpur, India; Turin, Italy; Parklands Suburb, Nairobi, Kenya;
Muscat, Oman; Oxford, UK, and Seattle, USA. Its primary aim was to study
growth, health, nutrition and neuro-development from less than 14 weeks of
gestation to 2 years of age, using the same conceptual framework as the WHO
Multicentre Growth Reference Study, so as to produce international
prescriptive standards for pregnancy dating, maternal weight gain, fetal
growth, newborn size, the postnatal growth of preterm infants and cognitive
development at 2 years of age. These scientifically robust clinical tools
may be used to monitor and evaluate maternal and fetal well-being, as well as
infant health and nutrition at an individual and population level.}
\description{
\if{html}{\figure{logo.png}{options: style='float: right' alt='logo' width='120'}}

The INTERGROWTH-21st Project is a multi-centre, multi-ethnic, population-based project, conducted between 2009 and 2014, in eight demarcated urban areas: Pelotas, Brazil; Shunyi County, Beijing, China; Central Nagpur, India; Turin, Italy; Parklands Suburb, Nairobi, Kenya; Muscat, Oman; Oxford, UK, and Seattle, USA. Its primary aim was to study growth, health, nutrition and neuro-development from less than 14 weeks of gestation to 2 years of age, using the same conceptual framework as the WHO Multicentre Growth Reference Study, so as to produce international prescriptive standards for pregnancy dating, maternal weight gain, fetal growth, newborn size, the postnatal growth of preterm infants and cognitive development at 2 years of age. These scientifically robust clinical tools may be used to monitor and evaluate maternal and fetal well-being, as well as infant health and nutrition at an individual and population level.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://nutriverse.io/intergrowth/}
  \item \url{https://github.com/nutriverse/intergrowth}
  \item Report bugs at \url{https://github.com/nutriverse/intergrowth/issues}
}

}
\author{
\strong{Maintainer}: Ernest Guevarra \email{ernest@guevarra.io} (\href{https://orcid.org/0000-0002-4887-4415}{ORCID})

}
\keyword{internal}
