% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{sfh_gestage_centile}
\alias{sfh_gestage_centile}
\title{International symphysis-fundal height standards: table of centiles of
symphysis-fundal height based on gestation age in exact weeks}
\format{
A tibble with 189 rows and 3 columns:
\describe{
  \item{\code{week}}{Gestational age in exact weeks}
  \item{\code{centile}}{Centiles (3rd, 5th, 10th, 50th, 90th, 95th, 97th)}
  \item{\code{sfh}}{Symphysis-fundal height in centimetres}
}
}
\source{
Napolitano R, Dhami J, Ohuma EO, Ioannou C, Conde-Agudelo A,
    Kennedy SH et al. Pregnancy dating by fetal crown-rump length: a
    systematic review of charts. BJOG 2014, 121(5):556-5.
}
\usage{
sfh_gestage_centile
}
\description{
International symphysis-fundal height standards: table of centiles of
symphysis-fundal height based on gestation age in exact weeks
}
\examples{
sfh_gestage_centile

}
\keyword{datasets}
