% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{sfh_gestage_sd}
\alias{sfh_gestage_sd}
\title{International symphysis-fundal height standards: table of z-scores of
symphysis-fundal height based on gestation age in exact weeks}
\format{
A tibble with 189 rows and 3 columns:
\describe{
  \item{\code{gestage}}{Gestational age in exact weeks}
  \item{\code{sd}}{z-scores (-3SD, -2SD, -1SD, 0, 1SD, 2SD, 3SD)}
  \item{\code{sfh}}{Symphysis-fundal height in centimetres}
}
}
\source{
Papageorghiou AT, Ohuma EO, Gravett MG, Lamber A, Noble JA,
  Pang R et al. International standards for symphysis-fundal height based
  on serial measurements from the Fetal Growth Longitudinal Study of the
  INTERGROWTH-21st Project: prospective cohort study in eight countries.
  BMJ 2016, 355:i5662.
}
\usage{
sfh_gestage_sd
}
\description{
International symphysis-fundal height standards: table of z-scores of
symphysis-fundal height based on gestation age in exact weeks
}
\examples{
sfh_gestage_sd

}
\keyword{datasets}
