% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{wlr_gestage_term_sd}
\alias{wlr_gestage_term_sd}
\title{International Newborn Size Standards: table of z-score for
weight-length ratio-for-gestational age for infants}
\format{
A tibble with 980 rows and 5 columns:
\describe{
\item{\code{week}}{Gestational age weeks component}
\item{\code{day}}{Gestational age days component}
\item{\code{sex}}{Sex; 1 = male; 2 = female}
\item{\code{sd}}{Standard deviation (z-score): -3, -2, -1, 0, 1, 2, 3}
\item{\code{wlr}}{Weight-length ratio in kilogram/metre}
}
}
\source{
Villar J, Giuliani F, Fenton TR, Ohuma EO, Ismail LC,
Kennedy SH et al. INTERGROWTH-21st very preterm size at birth reference
charts. Lancet 2016, 387(10021):844-45.
}
\usage{
wlr_gestage_term_sd
}
\description{
International Newborn Size Standards: table of z-score for
weight-length ratio-for-gestational age for infants
}
\examples{
wlr_gestage_term_sd

}
\keyword{datasets}
