% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{sfh_gestage_centile}
\alias{sfh_gestage_centile}
\title{International symphysis-fundal height standards: table of centiles of
symphysis-fundal height based on gestation age in exact weeks}
\format{
A tibble with 189 rows and 3 columns:
\describe{
  \item{\code{week}}{Gestational age in exact weeks}
  \item{\code{centile}}{Centiles (3rd, 5th, 10th, 50th, 90th, 95th, 97th)}
  \item{\code{sfh}}{Symphysis-fundal height in centimetres}
}
}
\source{
Papageorghiou AT, Ohuma EO, Gravett MG, Lamber A, Noble JA,
  Pang R et al. International standards for symphysis-fundal height based
  on serial measurements from the Fetal Growth Longitudinal Study of the
  INTERGROWTH-21st Project: prospective cohort study in eight countries.
  BMJ 2016, 355:i5662.
}
\usage{
sfh_gestage_centile
}
\description{
International symphysis-fundal height standards: table of centiles of
symphysis-fundal height based on gestation age in exact weeks
}
\examples{
sfh_gestage_centile

}
\keyword{datasets}
